restructure plugins into data, metrics, models, benchmarks (#50)
* restructure huggingface helper

* restructure Futrell2018

* remove __init__.py files to discourage direct imports

* restructure pereira2018

* restructure wikitext, fedorenko2016, and blank2014

* rename plugin pools to `<type>_registry`

`datasets` -> `data_registry`
`metrics` -> `metric_registry`
`benchmarks` -> `benchmark_registry`
`models` -> `model_registry`

* update plugin runner for directory restructure  (#52)

* update plugin paths

* typo fix

* add bibtex

* unique namespace for plugins

* nix relative import

* update plugin name

* Update brainscore_language/benchmarks/pereira2018/benchmark.py

* Update brainscore_language/utils/plugins/test_plugins.py

* update plugin tests path

* add `models`


* uplevel plugins

* reflect dir restructuring

* update plugins test instructions

* update paths


* rename {data,metric,benchmark}.py --> __init__.py

* rename plugins -> plugin_management

otherwise users might expect that the plugins/ folder contains the actual plugins

* autoformat; fix `self.plugin_env_path`

* use `bash` to `run_plugin.sh`

to avoid PermissionDenied error

* nitty oxford comma
