+++ brainscore_language/data/wikitext/__init__.py
@@ -1,0 +1,12 @@
+from datasets import load_dataset
+
+from brainscore_language import data_registry
+
+
+def wikitext2TestFromHuggingface():
+    dataset = load_dataset('wikitext', 'wikitext-2-raw-v1', split='test')
+    lines = dataset['text']
+    return lines
+
+
+data_registry['wikitext-2/test'] = wikitext2TestFromHuggingface
