+++ tests/test_test_plugins.py
@@ -4,10 +4,11 @@ import subprocess
 import textwrap
 from pathlib import Path
 
-from brainscore_language.plugins.test_plugins import PluginTestRunner
+from brainscore_language.plugin_management.test_plugins import PluginTestRunner
 
 DUMMY_PLUGIN = "dummy_plugin"
 DUMMY_PLUGIN_PATH = Path(__file__).parent / DUMMY_PLUGIN
+DUMMY_TYPE = Path(__file__).parent.name
 DUMMY_TESTFILE = DUMMY_PLUGIN_PATH / "test.py"
 DUMMY_REQUIREMENTS = DUMMY_PLUGIN_PATH / "requirements.txt"
 
@@ -28,7 +29,7 @@ class TestPluginTestRunner:
 
 	def test_plugin_name(self):
 		plugin_test_runner = PluginTestRunner(DUMMY_PLUGIN_PATH)
-		assert plugin_test_runner.plugin_name == DUMMY_PLUGIN
+		assert plugin_test_runner.plugin_name == DUMMY_TYPE + '_' + DUMMY_PLUGIN
 		
 	def test_get_conda_base(self):
 		plugin_test_runner = PluginTestRunner(DUMMY_PLUGIN_PATH)
